+++ datumaro/plugins/kitti_raw_format/extractor.py
@@ -14,6 +14,7 @@ from datumaro.components.extractor import DatasetItem, Importer, SourceExtractor
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.util import cast
 from datumaro.util.image import find_images
+from datumaro.util.meta_file_util import has_meta_file, parse_meta_file
 
 from .format import KittiRawPath, OcclusionStates, TruncationStates
 
@@ -136,11 +137,16 @@ class KittiRawExtractor(SourceExtractor):
 
         special_attrs = KittiRawPath.SPECIAL_ATTRS
         common_attrs = ['occluded']
-        label_cat = LabelCategories(attributes=common_attrs)
-        for label, attrs in sorted(labels.items(), key=lambda e: e[0]):
-            label_cat.add(label, attributes=set(attrs) - special_attrs)
 
-        categories = {AnnotationType.label: label_cat}
+        if has_meta_file(path):
+            categories =  { AnnotationType.label: LabelCategories.
+                from_iterable(parse_meta_file(path).keys()) }
+        else:
+            label_cat = LabelCategories(attributes=common_attrs)
+            for label, attrs in sorted(labels.items(), key=lambda e: e[0]):
+                label_cat.add(label, attributes=set(attrs) - special_attrs)
+
+            categories = {AnnotationType.label: label_cat}
 
         items = {}
         for idx, track in enumerate(tracks):