+++ CHANGELOG.md
@@ -8,18 +8,40 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## \[Unreleased\]
 ### Added
+- Video reading API
+  (<https://github.com/openvinotoolkit/datumaro/pull/521>)
 - Mapillary Vistas dataset format (Import-only)
   (<https://github.com/openvinotoolkit/datumaro/pull/537>)
 - Datumaro can now be installed on Windows on Python 3.9
   (<https://github.com/openvinotoolkit/datumaro/pull/547>)
+- Import for SYNTHIA dataset format
+  (<https://github.com/openvinotoolkit/datumaro/pull/532>)
+- Support of `score` attribute in KITTI detetion
+  (<https://github.com/openvinotoolkit/datumaro/pull/571>)
+- Support for Accuracy Checker dataset meta files in formats
+  (<https://github.com/openvinotoolkit/datumaro/pull/553>,
+  <https://github.com/openvinotoolkit/datumaro/pull/569>,
+  <https://github.com/openvinotoolkit/datumaro/pull/575>)
+- Import for VoTT dataset format
+  (<https://github.com/openvinotoolkit/datumaro/pull/573>)
+- Image resizing transform
+  (<https://github.com/openvinotoolkit/datumaro/pull/581>)
 
 ### Changed
 - The following formats can now be detected unambiguously:
-  `ade20k2017`, `ade20k2020`, `cvat`, `datumaro`, `icdar_text_localization`,
-  `icdar_text_segmentation`, `icdar_word_recognition`, `kitti_raw`, `label_me`,
-  `mot_seq`, `yolo`
+  `ade20k2017`, `ade20k2020`, `camvid`, `coco`, `cvat`, `datumaro`,
+  `icdar_text_localization`, `icdar_text_segmentation`,
+  `icdar_word_recognition`, `imagenet_txt`, `kitti_raw`, `label_me`, `lfw`,
+  `mot_seq`, `open_images`, `vgg_face2`, `voc`, `widerface`, `yolo`
   (<https://github.com/openvinotoolkit/datumaro/pull/531>,
-  <https://github.com/openvinotoolkit/datumaro/pull/536>)
+  <https://github.com/openvinotoolkit/datumaro/pull/536>,
+  <https://github.com/openvinotoolkit/datumaro/pull/550>,
+  <https://github.com/openvinotoolkit/datumaro/pull/557>,
+  <https://github.com/openvinotoolkit/datumaro/pull/558>)
+- Allowed Pytest-native tests
+  (<https://github.com/openvinotoolkit/datumaro/pull/563>)
+- Allowed export options in the `datum merge` command
+  (<https://github.com/openvinotoolkit/datumaro/pull/545>)
 
 ### Deprecated
 - Using `Image`, `ByteImage` from `datumaro.util.image` - these classes
@@ -27,10 +49,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/538>)
 
 ### Removed
-- TBD
+- Equality comparison support between `datumaro.components.media.Image`
+  and `numpy.ndarray`
+  (<https://github.com/openvinotoolkit/datumaro/pull/568>)
 
 ### Fixed
-- TBD
+- Bug #560: import issue with MOT dataset when using seqinfo.ini file
+  (<https://github.com/openvinotoolkit/datumaro/pull/564>)
+- Empty lines in VOC subset lists are not ignored
+  (<https://github.com/openvinotoolkit/datumaro/pull/587>)
 
 ### Security
 - TBD
@@ -64,7 +91,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Unhandled exception when a file is specified as the source for a COCO or
   MOTS dataset
   (<https://github.com/openvinotoolkit/datumaro/pull/530>)
-
+- Exporting dataset without `color` attribute into the
+  `icdar_text_segmentation` format
+  (<https://github.com/openvinotoolkit/datumaro/pull/556>)
 ### Security
 - TBD